requirements.txt
flask
requests
python-dotenv
sqlalchemy
sqlalchemy_serializer